+++ modules/controllers/01-network.tf
@@ -13,8 +13,8 @@ resource "aws_subnet" "k8s_private" {
 
   tags = {
     Environment                       = var.environment
-    ManagedBy                         = "terraform (k8s module)"
-    ModuleRepository                  = "https://github.com/jecnua/terraform-aws-kubernetes"
+    ManagedBy                         = "opentofu (k8s module)"
+    ModuleRepository                  = "https://github.com/jecnua/opentofu-aws-kubernetes"
     Name                              = "${var.unique_identifier} ${var.environment} k8s private subnet"
     KubernetesCluster                 = var.kubernetes_cluster
     "kubernetes.io/role/internal-elb" = "true" # Needed for integration with load balancing
@@ -30,8 +30,8 @@ resource "aws_subnet" "k8s_public" {
 
   tags = {
     Environment              = var.environment
-    ManagedBy                = "terraform (k8s module)"
-    ModuleRepository         = "https://github.com/jecnua/terraform-aws-kubernetes"
+    ManagedBy                = "opentofu (k8s module)"
+    ModuleRepository         = "https://github.com/jecnua/opentofu-aws-kubernetes"
     Name                     = "${var.unique_identifier} ${var.environment} k8s public subnet"
     KubernetesCluster        = var.kubernetes_cluster
     "kubernetes.io/role/elb" = "true" # Needed for integration with load balancing
@@ -43,8 +43,8 @@ resource "aws_route_table" "k8s_private_route_table" {
 
   tags = {
     Environment       = var.environment
-    ManagedBy         = "terraform (k8s module)"
-    ModuleRepository  = "https://github.com/jecnua/terraform-aws-kubernetes"
+    ManagedBy         = "opentofu (k8s module)"
+    ModuleRepository  = "https://github.com/jecnua/opentofu-aws-kubernetes"
     Name              = "${var.unique_identifier} ${var.environment} k8s private route table"
     KubernetesCluster = var.kubernetes_cluster
   }
@@ -55,8 +55,8 @@ resource "aws_route_table" "k8s_public_route_table" {
 
   tags = {
     Environment       = var.environment
-    ManagedBy         = "terraform (k8s module)"
-    ModuleRepository  = "https://github.com/jecnua/terraform-aws-kubernetes"
+    ManagedBy         = "opentofu (k8s module)"
+    ModuleRepository  = "https://github.com/jecnua/opentofu-aws-kubernetes"
     Name              = "${var.unique_identifier} ${var.environment} k8s public route table"
     KubernetesCluster = var.kubernetes_cluster
   }