+++ modules/controllers/02-locals.tf
@@ -4,8 +4,8 @@ locals {
       "Environment"        = format("%s", var.environment)
       "k8s.io/role/master" = "1" # Taken from the kops # TODO: CHECK
       "KubernetesCluster"  = var.kubernetes_cluster
-      "ManagedBy"          = "terraform k8s module"
-      "ModuleRepository"   = "https://github.com/jecnua/terraform-aws-kubernetes"
+      "ManagedBy"          = "opentofu k8s module"
+      "ModuleRepository"   = "https://github.com/jecnua/opentofu-aws-kubernetes"
     },
     var.additional_tags_as_map,
   )