+++ modules/cri/cri-o/00-variables_defaults.tf
@@ -1,5 +1,5 @@
-variable "crio_version" {
+variable "kubernetes_version" {
   type        = string
   description = "Version to install"
-  default     = "1.25"
+  default     = "1.29"
 }