+++ modules/cri/CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+## 9.0.0
+
+### Breaking changes
+
+### Features & Changes
+
+- Updated cri-o version to latest
+- Updated cri-o installation method
+- Generic cleanup of cri-o installation script
+
+### Bugfixes
+
+### Known bugs/issues
+
 ## 6.0.0
 
 ### Breaking changes